+++ CHANGELOG.md
@@ -3,6 +3,7 @@ A list of ongoing changes per version release.
 
 ## V0.1.3
 - Added numista API support for searching numista number which will then populate `Coin_Types.csv`
+- Users can now change country/denomination through a single session, if type you found in unowned, it prompts you to enter a numista number which will use the API to gather the Country, Denomination, and year range for that coin.
 
 ## V0.1.2
 - Removed `Collection Sorting` and `Collection Audit` session types.